requirements.txt
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
deprecation==2.1.0
Django==4.2.5
exceptiongroup==1.1.3
gotrue==1.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
packaging==23.1
postgrest==0.10.8
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
typing_extensions==4.7.1
websockets==10.4
